refactor: PathRepository 를 PathFinder 도메인 객체로 리팩토링
