+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -1,7 +1,13 @@
 package nextstep.subway.path.application;
 
+import java.util.Collection;
+import java.util.List;
+import java.util.stream.Collectors;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.line.domain.Section;
 import nextstep.subway.path.domain.Path;
-import nextstep.subway.path.domain.PathRepository;
+import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.path.exception.SourceTargetNotFoundException;
 import nextstep.subway.station.domain.Station;
@@ -10,23 +16,32 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class PathService {
-    private final PathRepository pathRepository;
     private final StationRepository stationRepository;
+    private final LineRepository lineRepository;
 
-    public PathService(PathRepository pathRepository, StationRepository stationRepository) {
-        this.pathRepository = pathRepository;
+    public PathService(StationRepository stationRepository, LineRepository lineRepository) {
         this.stationRepository = stationRepository;
+        this.lineRepository = lineRepository;
     }
 
     public PathResponse findShortestPath(Long source, Long target) {
         Station sourceStation = findStationById(source);
         Station targetStation = findStationById(target);
-        Path path = pathRepository.findShortestPath(sourceStation, targetStation);
+        Path path = new PathFinder(findAllSections())
+                .findShortestPath(sourceStation, targetStation);
         return PathResponse.of(path);
     }
 
     private Station findStationById(Long stationId) {
         return stationRepository.findById(stationId)
                 .orElseThrow(SourceTargetNotFoundException::new);
+    }
+
+    private List<Section> findAllSections() {
+        return lineRepository.findAll()
+                .stream()
+                .map(Line::getSections)
+                .flatMap(Collection::stream)
+                .collect(Collectors.toList());
     }
 }
