+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -3,9 +3,6 @@ package nextstep.subway.path.domain;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
-import java.util.stream.Collectors;
-import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.domain.Section;
 import nextstep.subway.path.exception.InvalidSourceTargetException;
 import nextstep.subway.path.exception.PathNotFoundException;
@@ -14,17 +11,14 @@ import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
-import org.springframework.stereotype.Component;
 
-@Component
-public class DijkstraPathRepository implements PathRepository {
-    private final LineRepository lineRepository;
+public class PathFinder {
+    private final List<Section> sections;
 
-    public DijkstraPathRepository(LineRepository lineRepository) {
-        this.lineRepository = lineRepository;
+    public PathFinder(List<Section> sections) {
+        this.sections = sections;
     }
 
-    @Override
     public Path findShortestPath(Station source, Station target) {
         if (source.equals(target)) {
             throw new InvalidSourceTargetException();
@@ -37,7 +31,6 @@ public class DijkstraPathRepository implements PathRepository {
     }
 
     private DijkstraShortestPath<Station, DefaultWeightedEdge> createDijkstra() {
-        List<Section> sections = findAllSections();
         WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph<>(DefaultWeightedEdge.class);
         addVertex(graph, sections);
         setEdgeWeight(graph, sections);
@@ -57,13 +50,5 @@ public class DijkstraPathRepository implements PathRepository {
             int weight = section.getDistance();
             graph.setEdgeWeight(edge, weight);
         }
-    }
-
-    private List<Section> findAllSections() {
-        return lineRepository.findAll()
-                .stream()
-                .map(Line::getSections)
-                .flatMap(Collection::stream)
-                .collect(Collectors.toList());
     }
 }
