--- src/main/java/nextstep/subway/path/domain/PathRepository.java
@@ -1,7 +1,0 @@
-package nextstep.subway.path.domain;
-
-import nextstep.subway.station.domain.Station;
-
-public interface PathRepository {
-    Path findShortestPath(Station source, Station target);
-}
